@@ -2,7 +2,7 @@
 const nextConfig = {
   reactStrictMode: true,
   images: {
-    domains: ['source.unsplash.com'],
+    domains: ['picsum.photos'],
   },
 }
 
